Re-write event message service
Added Message, Query Handling to message service. Also in send messages will use message service.
